requirements.txt
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
opencv-python==4.12.0.88
numpy==2.2.6
matplotlib==3.10.7
pillow==11.3.0
tqdm==4.67.1